==> Scripts/03 PDF. SNAP Data Maps.qmd <==
---
title: "Overview of SNAP in Illinois"
author: "Caroline Davidson"
format: pdf
---



```{r setup}
#| echo: false
#| include: false
knitr::opts_chunk$set(echo = FALSE)

knitr::opts_knit$set(root.dir = rprojroot::find_rstudio_root_file()) # set main project folder as working directory

library(sf)
library(tidyverse)

```


```{r percentage households SNAP}
map_perc_hh_snap_county_all

```

## Initial Maps

```{r}
plot(county_acs["perc_hh_snap"], main = "Percentage of Households Receiving SNAP by County")

plot(tract_acs["perc_hh_snap"], main = "Percentage of Households Receiving SNAP by Census Tract")

plot(region_acs["perc_hh_snap"], main = "Percentage of Households Receiving SNAP by Region")

plot(region_acs["n_hh_snap"], main = "Number of Households Receiving SNAP by Region")

plot(region_acs["n_LinkUpPartners_all"], main = "Number of LinkUp Partners by Region")

```
